#18: move helper functions into utils folder && improve component structure
