+++ waste-lookup/src/components/FavoriteModal/FavoriteModal.css
